+++ ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: JFikents <JFikents@student.42Heilbronn.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 16:36:20 by JFikents          #+#    #+#             */
-/*   Updated: 2023/10/06 21:01:22 by JFikents         ###   ########.fr       */
+/*   Updated: 2023/10/08 17:17:31 by JFikents         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 
-void	*ft_memset(void *address, int filler, size_t size)
+void	*ft_memset(void *address, int filler, unsigned long size)
 {
 	if (size <= 0)
 		return (0);
